requirements.txt
tweepy==4.14.0
openai==1.3.0
python-dotenv==1.0.0
Pillow==10.0.0
requests==2.31.0
schedule==1.2.0
python-dateutil==2.8.2
tenacity==8.2.2 